+++ android/src/main/java/im/zego/zego_express_engine/internal/ZegoUtils.java
@@ -1,5 +1,8 @@
 package im.zego.zego_express_engine.internal;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+
 public class ZegoUtils {
     /* Utils */
 
@@ -21,5 +24,23 @@ public class ZegoUtils {
 
     public static double doubleValue(Number number) {
         return number != null ? number.doubleValue() : .0f;
+    }
+
+    public static int[] intArrayValue(ArrayList<Integer> integerArrayList) {
+        int[] ret = new int[integerArrayList.size()];
+        Iterator<Integer> iterator = integerArrayList.iterator();
+        for (int i = 0; i < ret.length; i++) {
+            ret[i] = iterator.next();
+        }
+        return ret;
+    }
+
+    public static float[] floatArrayValueFromDoubleArray(ArrayList<Double> floatArrayList) {
+        float[] ret = new float[floatArrayList.size()];
+        Iterator<Double> iterator = floatArrayList.iterator();
+        for (int i = 0; i < ret.length; i++) {
+            ret[i] = iterator.next().floatValue();
+        }
+        return ret;
     }
 }